GCS pipeline user guide first draft
